% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaAnalysis.R
\name{doMetaAnalysis}
\alias{doMetaAnalysis}
\title{Create figures and tables for report}
\usage{
doMetaAnalysis(outputFolders, maOutputFolder, maxCores)
}
\arguments{
\item{outputFolders}{Vector of names of local folders where the results were generated; make sure 
to use forward slashes (/). D}

\item{maOutputFolder}{A local folder where the meta-anlysis results will be written.}

\item{maxCores}{How many parallel cores should be used? If more cores are made available
this can speed up the analyses.}
}
\description{
Create figures and tables for report
}
\details{
This function generates tables and figures for the report on the study results.
}
